refactor: real fake ban.
